Move DTO-objects into database.dto-package
